+++ hosts/matangi/default.nix
@@ -173,6 +173,7 @@
         enable = true;
         motherboard = "intel";
       };
+      lenovo.enable = true;
       sound.enable = true;
       networking = {
         enable = true;